Backport fix for duplicate embedded document creation
Mongoid's legacy branch has a problem where some embedded documents will
get created twice.  This is caused by a failure to clear new_record
during an initial insert, causing the embedded document to be later
saved a second time when update_embedded is called during an top-level
document save.

The following two patches on the master branch seem to fix this problem
(and some others besides):

http://github.com/durran/mongoid/commit/4fd1168ec6a75361db527855b05ce70ffe87ca41
http://github.com/durran/mongoid/commit/334c299f856057bf3e6a9f359cf7200dab340b6b

This patch backports these changes (minus the update_attributes changes)
and it appears to fix the problem without breaking any of the existing
specs.