Allow referenced_in association to be nil
Until a few days ago, Mongoid supported objects which were optionally
referenced_in a parent object's references_many, but which could also
stand on their own.  Recently, this has stopped working, and started
generating errors about an object with an id of 4.

    Failure/Error: @account.creator.should be_nil
    Document not found for class User with id(s) 4
    # ./spec/../lib/mongoid/contexts/ids.rb:19:in `id_criteria'
    # ./spec/../lib/mongoid/criteria.rb:31:in `__send__'
    # ./spec/../lib/mongoid/criteria.rb:31:in `id_criteria'
    # ./spec/../lib/mongoid/criteria.rb:189:in `translate'
    # ./spec/../lib/mongoid/finders.rb:68:in `find'
    # ./spec/../lib/mongoid/associations/referenced_in.rb:21:in `initialize'
    # ./spec/../lib/mongoid/associations.rb:280:in `new'
    # ./spec/../lib/mongoid/associations.rb:280:in `creator'
    # ./spec/../lib/mongoid/memoization.rb:11:in `memoized'
    # ./spec/../lib/mongoid/associations.rb:279:in `creator'
    # ./spec/integration/mongoid/associations_spec.rb:224

These errors are caused by the following two lines of code in
Association's update_foreign_keys method:

    target = send(name)
    send("#{foreign_key}=", target ? target.id : nil)

This code is manipulating proxy objects, which don't do the expected
thing when evaluated in a conditional statement:

    class ExampleProxy < Mongoid::Associations::Proxy
      def initialize(target)
        @target = target
      end
    end

    proxy = ExampleProxy.new(nil)
    proxy.nil? => false
    (proxy ? true : false) => true

According to folks on the ruby IRC channel, there's no way to make a
proxy object reliably behave like nil when used in a conditional
expression.

If we make the use of proxies explicit in this code, then all the unit
tests pass and we stop accidentally storing 4 (the id of nil!) in the
database.

    proxy = send(name)
    send("#{foreign_key}=", proxy && proxy.target ? proxy.id : nil)

There's an analogous use of proxy.target in the associate method:

    proxy.target ? proxy : nil

...so this behavior is consistent with other code in Mongoid.